+++ packages/extension-chakra-storefront/jest.config.js
@@ -60,7 +60,7 @@ module.exports = {
         '<rootDir>/src/pages/checkout/partials/contact-info.test.js',
         '<rootDir>/src/pages/checkout/partials/login-state.test.js',
         '<rootDir>/src/utils/responsive-image.test.js',
-        '<rootDir>/src/utils/error-utils.test.js',
+        '<rootDir>/src/hooks/use-errors.test.js',
         '<rootDir>/src/hooks/use-toast.test.js',
         '<rootDir>/src/pages/product-detail/metadata.test.js',
         '<rootDir>/src/pages/product-list/metadata.test.js',
