+++ packages/extension-chakra-storefront/src/pages/checkout/partials/payment.jsx
@@ -22,7 +22,7 @@ import PaymentForm from '../../../pages/checkout/partials/payment-form'
 import ShippingAddressSelection from '../../../pages/checkout/partials/shipping-address-selection'
 import AddressDisplay from '../../../components/address-display'
 import {PromoCode, usePromoCode} from '../../../components/promo-code'
-import {useErrorHandler} from '../../../utils/error-utils'
+import {useErrorHandler} from '../../../hooks/use-errors'
 
 const Payment = () => {
     const {formatMessage} = useIntl()
